+++ packages/terra-framework-docs/src/terra-dev-site/test/slider/SliderWithOnChangeCallback.test.jsx
@@ -1,0 +1,26 @@
+import React, { useState } from 'react';
+import Slider from 'terra-slider';
+
+const SliderWithOnChangeCallback = () => {
+  const [value, setValue] = useState('20');
+
+  const handleChange = (v) => {
+    setValue(v);
+  };
+  const message = `Slider value is: ${value}`;
+
+  return (
+    <div>
+      <Slider
+        minimumValue={0}
+        maximumValue={100}
+        label="Zoom"
+        value={value}
+        onChange={handleChange}
+      />
+      <p>{message}</p>
+    </div>
+  );
+};
+
+export default SliderWithOnChangeCallback;