+++ packages/terra-framework-docs/CHANGELOG.md
@@ -9,14 +9,14 @@
 
 * Fixed
   * Fixed the `terra-tabs` icon-only examples keyboard navigation issue.
+  * Fixed broken links in the Worklist Data Grid documentation in `terra-data-grid` component.
 
 * Added
   * Added test for data grid column resizing.
   * Added example for Flowsheet Data Grid containing "No Result" cells.
 
-* Fixed
-  * Fixed the `terra-tabs` icon-only examples keyboard navigation issue.
-  * Fixed broken links in the Worklist Data Grid documentation in `terra-data-grid` component.
+* Updated
+  * Updated `terra-table` example to show providing custom masked cell message label.
 
 ## 1.42.0 - (October 20, 2023)